+++ packages/api/src/platforms/vtex/clients/search/index.ts
@@ -1,9 +1,12 @@
-import type { Context, Options } from '../../index'
-import { fetchAPI } from '../fetch'
-import type { SelectedFacet } from '../../utils/facets'
-import type { ProductSearchResult } from './types/ProductSearchResult'
-import type { FacetSearchResult } from './types/FacetSearchResult'
 import type { IStoreSelectedFacet } from '../../../../__generated__/schema'
+import type { Context, Options } from '../../index'
+import type { SelectedFacet } from '../../utils/facets'
+import { fetchAPI } from '../fetch'
+import type { FacetSearchResult } from './types/FacetSearchResult'
+import type {
+  ProductSearchResult,
+  Suggestion,
+} from './types/ProductSearchResult'
 
 export type Sort =
   | 'price:desc'
@@ -84,11 +87,27 @@ export const IntelligentSearch = (
   const products = (args: Omit<SearchArgs, 'type'>) =>
     search<ProductSearchResult>({ ...args, type: 'product_search' })
 
+  const suggestedProducts = (
+    args: Omit<SearchArgs, 'type'>
+  ): Promise<ProductSearchResult> =>
+    fetchAPI(
+      `${base}/_v/api/intelligent-search/product_search?query=${args.query}`
+    )
+
+  const suggestedTerms = (
+    args: Omit<SearchArgs, 'type'>
+  ): Promise<Suggestion> =>
+    fetchAPI(
+      `${base}/_v/api/intelligent-search/search_suggestions?query=${args.query}`
+    )
+
   const facets = (args: Omit<SearchArgs, 'type'>) =>
     search<FacetSearchResult>({ ...args, type: 'facets' })
 
   return {
     facets,
     products,
+    suggestedTerms,
+    suggestedProducts,
   }
 }