+++ packages/api/src/platforms/vtex/resolvers/productGroup.ts
@@ -5,24 +5,24 @@ import type { StoreProduct } from './product'
 
 type Root = PromiseType<ReturnType<typeof StoreProduct.isVariantOf>>
 
-const BLOCKED_PROPERTIES: Record<string, boolean> = {
-  sellerId: true,
-}
+const BLOCKED_SPECIFICATIONS = new Set(['allSpecifications'])
 
 export const StoreProductGroup: Record<string, Resolver<Root>> = {
   hasVariant: (root) =>
     root.isVariantOf.items.map((item) => enhanceSku(item, root.isVariantOf)),
   productGroupID: ({ isVariantOf }) => isVariantOf.productId,
   name: ({ isVariantOf }) => isVariantOf.productName,
-  additionalProperty: ({ isVariantOf: { properties } }) =>
-    properties.flatMap((property) => {
-      if (BLOCKED_PROPERTIES[property.name]) {
-        return []
-      }
-
-      return property.values.map((propertyValue) => ({
-        name: property.name,
-        value: propertyValue,
-      }))
-    }),
+  additionalProperty: ({ isVariantOf: { specificationGroups } }) =>
+    specificationGroups
+      // filter sku specifications so we dont mess sku with product specs
+      .filter(
+        (specificationGroup) =>
+          !BLOCKED_SPECIFICATIONS.has(specificationGroup.name)
+      )
+      // Transform specs back into product specs
+      .flatMap(({ specifications }) =>
+        specifications.flatMap(({ name, values }) =>
+          values.map((value) => ({ name, value }))
+        )
+      ),
 }